Normal Booting and Safe booting
