+++ firmware/safe_boot.py
@@ -1,4 +1,4 @@
-"""main.py"""
+"""safe_boot.py"""
 from machine import Pin, UART, PWM, I2C
 import actuator_constants as actuator_constants
 import movements
@@ -18,11 +18,8 @@ song4 = ["A5", "A5", "E5", "E5"]
 
 SLEEP_COUNT = 0
 
-try:
-    i2c = I2C(0, sda=Pin(8), scl=Pin(9), freq=400000)
-    oled = SSD1306_I2C(128, 64, i2c)
-except:
-    pass
+#i2c = I2C(0, sda=Pin(8), scl=Pin(9), freq=400000)
+#oled = SSD1306_I2C(128, 64, i2c)
 
 TH = bytearray(
     b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\xf0\x00\x00\x00\x00\x00\x00\x03\xff\xd0\x00\x00\x00\x00\x0f\xff\xff\x00\x00\x00\x00\x00\x00\x1f\xff\xfc\x00\x00\x00\x00?\xff\xff\xe0\x00\x00\x00\x00\x00\xff\xff\xff\x80\x00\x00\x00\xff\x80\x0b\xf8\x00\x00\x00\x00\x03\xf4\x01\x7f\xe0\x00\x00\x03\xf0\x00\x00\x0e\x00\x00\x00\x00\x0e\x00\x00\x03\xf8\x00\x00\x07\xc0\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00|\x00\x00\x1e\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x0f\x00\x000\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x80\x00@\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xc0\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x17\xe8\x00\x00\x00\x00\x00\x00\x02\xfa\x00\x00\x00\x00\x00\x01\xfe\xff\x80\x00\x00\x00\x00\x00?\xaf\xf0\x00\x00\x00\x00\x0f\x80\x07\xf0\x00\x00\x00\x00\x03\xf8\x00~\x00\x00\x00\x00|\x00\x00~\x00\x00\x00\x00\x0f\xc0\x00\x07\x80\x00\x00\x01\xf0\x00\x00\x1f\x00\x00\x00\x00>\x00\x00\x01\xe0\x00\x00\x03\xc0\x00\x00\x07\xc0\x00\x00\x00x\x00\x00\x00\xf8\x00\x00\x0f\x80\x00\x00\x01\xf0\x00\x00\x01\xf0\x00\x00\x00<\x00\x00\x0f\x00\x00\x00\x00\xf8\x00\x00\x03\xe0\x00\x00\x00\x1f\x00\x00>\x00\x00\x00\x00|\x00\x00\x07\xc0\x00\x00\x00\x0f\x00\x00<\x00\x00\x00\x00<\x00\x00\x0f\x80\x00\x00\x00\x07\xc0\x00x\x00\x00\x00\x00\x1f\x00\x00\x1f\x00\x00\x00\x00\x03\xc0\x00\xf0\x00\x00\x00\x00\x1f\x00\x00\x1e\x00\x00\x00\x00\x01\xe0\x01\xf0\x00\x00\x00\x00\x0f\x80\x00>\x00\x00\x00\x00\x01\xf0\x01\xe0\x00\x00\x00\x00\x07\xc0\x00|\x00\x00\x00\x00\x01\xf0\x03\xe0\x00\x00\x00\x00\x07\xc0\x00|\x00\x00\x00\x00\x00\xf0\x03\xe0\x00\x00\x00\x00\x07\xc0\x00|\x00\x00\x00\x00\x00\xf8\x03\xe0\x00\x00\x00\x00\x03\xe0\x00\xf8\x00\x00\x00\x00\x00\xf8\x07\xc0\x00\x00\x00\x00\x03\xe0\x00\xf8\x00\x00\x00\x00\x00|\x07\xc0\x00\x00\x00\x00\x03\xf0\x01\xf8\x00\x00\x00\x00\x00\xfc\x07\xe0\x00\x00\x00\x00\x03\xe0\x00\xf8\x00\x00\x00\x00\x00|\x07\xc0\x00\x00\x00\x00\x03\xf0\x01\xf8\x00\x00\x00\x00\x00\xfc\x07\xe0\x00\x07\xfe\x00\x03\xf0\x01\xf8\x00\x1f\xfc\x00\x00\xfc\x07\xe0\x009\x7f\xc0\x03\xf0\x01\xf8\x00\xff\xa7\x80\x00\xf8\x03\xe0\x00\xe0\x0f\xf0\x07\xe0\x01\xf8\x03\xfe\x00\xe0\x00\xfc\x07\xf0\x01\xc0\x07\xfc\x03\xf0\x00\xfc\x07\xfc\x00p\x01\xf8\x03\xf0\x03\xc0\x07\xfe\x07\xe0\x00\xfc\x0f\xfc\x00x\x01\xf8\x03\xf0\x07\xe0\x1f\xfe\x0f\xe0\x00\xfc\x0f\xff\x00\xfc\x03\xf8\x01\xf8\x07\xfe\xff\xff\x0f\xc0\x00\xfe\x1f\xff\xef\xfc\x03\xf0\x01\xfc\x0f\xff\xff\xff\x1f\xc0\x00\x7f\x1f\xff\xff\xfc\x07\xe0\x00\xfe\x0f\xff\xff\xff\x1f\xc0\x00?\x1f\xff\xff\xfe\x0f\xe0\x00\x7f\x07\xff\xff\xff\x7f\x80\x00?\xdf\xff\xff\xfc\x1f\xc0\x00\x7f\x8f\xff\xff\xff\xff\x00\x00\x1f\xff\xff\xff\xfc?\xc0\x00?\xc7\xff\xff\xff\xfe\x00\x00\x1f\xff\xff\xff\xfc\xff\x80\x00\x1f\xf3\xff\xff\xff\xfe\x00\x00\x0f\xff\xff\xff\xf9\xff\x00\x00\x07\xff\xff\xff\xff\xf8\x00\x00\x03\xff\xff\xff\xff\xfc\x00\x00\x03\xff\xff\xff\xff\xf0\x00\x00\x03\xff\xff\xff\xff\xf8\x00\x00\x01\xff\xff\xff\xff\xe0\x00\x00\x00\xff\xff\xff\xff\xf0\x00\x00\x00\x7f\xff\xff\xff\x80\x00\x00\x00?\xff\xff\xff\xc0\x00\x00\x00\x1f\xff\xff\xfe\x00\x00\x00\x00\x1f\xff\xff\xff\x00\x00\x00\x00\x07\xff\xff\xf8\x00\x00\x00\x00\x03\xff\xff\xf8\x00\x00\x00\x00\x00\x7f\xff\x80\x00\x00\x00\x00\x00\x7f\xff\xa0\x00\x00\x00\x00\x00\x01H\x00\x00\x00\x00\x00\x00\x00\x90\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
@@ -231,3 +228,4 @@ if __name__ == "__main__":
                         for pos in range(9000, 7000, -50):
                             setServoCycle(pos)
                     movements.action_standup()
+
